requirements.txt
Django==3.2.16
django_taggit==3.1.0
